requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose
bcrypt